Converted base64 encode/decode to C++ and added coresponding base64URL decode/encode. Additionally added utility function to sign data.
Added methods to handle DPP connector, get TLS group ID, and method to split strings.

+++ src/em/prov/easyconnect/ec_util.cpp
@@ -193,7 +193,7 @@ std::pair<uint8_t*, uint16_t> ec_util::unwrap_wrapped_attrib(ec_attribute_t *wra
             printf("%s:%d: AAD input is NULL, AAD decryption failed!\n", __func__, __LINE__);
             return {nullptr, 0};
         }
-        size_t pre_wrapped_attribs_size = reinterpret_cast<uint8_t*>(wrapped_attrib) - frame_attribs;
+        size_t pre_wrapped_attribs_size = static_cast<size_t>(reinterpret_cast<uint8_t*>(wrapped_attrib) - frame_attribs);
         result = siv_decrypt(&ctx, wrapped_ciphertext, unwrap_attribs, wrapped_len,
                              wrapped_attrib->data, 2,
                              frame, frame_len,
